[MIG] l10n_ar_tax_ratio: integrate module in l10n_ar_tax
