+++ l10n_ar_tax/models/account_tax.py
@@ -1,5 +1,5 @@
 from odoo import api, fields, models
-from odoo.exceptions import UserError
+from odoo.exceptions import UserError, ValidationError
 
 
 class AccountTax(models.Model):
@@ -88,6 +88,7 @@ class AccountTax(models.Model):
         api_articulo_inciso_calculo_selection,
         string="Artículo/Inciso para el cálculo retención",
     )
+    ratio = fields.Float(default=100.00, help="Ratio to apply to tax base amount.")
 
     @api.ondelete(at_uninstall=False)
     def _check_tax_used_on_company_tax_fp(self):
@@ -96,3 +97,14 @@ class AccountTax(models.Model):
             raise UserError(
                 "Error se esta usando en ws de estas cias %s" % ws.mapped("fiscal_position_id.company_id.name")
             )
+
+    @api.constrains("ratio")
+    def _check_line_ids_percent(self):
+        """Check that the total percent is not bigger than 100.0"""
+        for tax in self:
+            if not tax.ratio or tax.ratio <= 0.0 or tax.ratio > 100.0:
+                raise ValidationError(
+                    self.env._(
+                        "The total percentage (%s) should be greater than 0 and less than or equal to 100.", tax.ratio
+                    )
+                )
