+++ l10n_ar_tax/views/account_tax_view.xml
@@ -18,6 +18,10 @@
                 <attribute name="invisible">l10n_ar_withholding_payment_type</attribute>
             </group>
 
+            <xpath expr="//field[@name='tax_group_id']" position="after">
+                <field name="ratio" string="Ratio %" invisible="country_code != 'AR' or l10n_ar_type_tax_use not in ('customer', 'supplier')"/>
+            </xpath>
+
             <!-- TODO mejorar esto, por ahora solo lo hacemos visible con este grupo ya que este grupo solo lo estamos usando en realidad para API, deberiamos ver de hacerlo de otra manera -->
             <!-- TODO agregar que sea de santa fe -->
             <!-- <i class="fa fa-exclamation-circle" data-tooltip="This field will not be taken into account for sales that come from eCommerce, in this case you must configure it directly in Settings" title="This field will not be taken into account for sales that come from eCommerce, in this case you must configure it directly in Settings" aria-hidden="true"/>  -->
