--- l10n_ar_tax_ratio/README.rst
@@ -1,61 +1,0 @@
-.. |company| replace:: ADHOC SA
-
-.. |company_logo| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-logo.png
-   :alt: ADHOC SA
-   :target: https://www.adhoc.com.ar
-
-.. |icon| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-icon.png
-
-.. image:: https://img.shields.io/badge/license-AGPL--3-blue.png
-   :target: https://www.gnu.org/licenses/agpl
-   :alt: License: AGPL-3
-
-=============================
-Argentinean Withholding Ratio
-=============================
-
-Este módulo implementa:
-
-* Permite establecer un ratio de 1 a 100 en el impuesto y representa el porcentaje de base imponible que se tendrá en cuenta para aplicar el impuesto, siempre y cuando el cálculo del impuesto sea de tipo porcentaje. Si bien se puede usar ese ratio para cualquier impuesto en principio fue solicitado solo para Córdoba.
-
-
-Installation
-============
-
-To install this module, you need to:
-
-#. Only need to install the module
-
-Configuration
-=============
-
-To configure this module, you need to:
-
-#. Nothing to configure
-
-Usage
-=====
-
-.. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
-   :alt: Try me on Runbot
-   :target: http://runbot.adhoc.com.ar/
-
-Credits
-=======
-
-Images
-------
-
-* |company| |icon|
-
-Contributors
-------------
-
-Maintainer
-----------
-
-|company_logo|
-
-This module is maintained by the |company|.
-
-To contribute to this module, please visit https://www.adhoc.com.ar.
