--- l10n_ar_tax_ratio/__init__.py
@@ -1,1 +1,0 @@
-from . import models
