--- l10n_ar_tax_ratio/__manifest__.py
@@ -1,18 +1,0 @@
-{
-    "name": "Argentinean Withholding Ratio",
-    "version": "18.0.1.0.0",
-    "category": "Localization/Argentina",
-    "sequence": 14,
-    "author": "ADHOC SA",
-    "website": "www.adhoc.com.ar",
-    "license": "AGPL-3",
-    "summary": "",
-    "depends": ["l10n_ar_tax"],
-    "data": [
-        "views/account_tax_view.xml",
-    ],
-    "demo": [],
-    "installable": False,
-    "auto_install": False,
-    "application": False,
-}
