--- l10n_ar_tax_ratio/models/__init__.py
@@ -1,6 +1,0 @@
-##############################################################################
-# For copyright and license notices, see __manifest__.py file in module root
-# directory
-##############################################################################
-from . import account_tax
-from . import l10n_ar_payment_withholding
